requirements.txt
matplotlib==3.3.0
numpy==1.19.0
pandas==1.0.5
Pillow==8.4.0
scikit_image==0.17.2
scipy==1.5.1
tensorboard==2.7.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.48.0
